# docs/swagger.yml
openapi: 3.0.0
info:
  title: movie-review-api
  version: 0.0.1
  description: movie-review-api
paths:
  /movies/recommended:
    get:
      summary: おすすめの映画リストを返却する
      responses:
        "200":
          description: 映画のリスト
          content:
            application/json:
              schema:
                type: object
                properties:
                  movies:
                    type: array
                    items:
                      $ref: "#/components/schemas/MovieSummary"
  /movies/search:
    get:
      summary: 映画の検索結果を返却する
      parameters:
        - name: query
          in: query
          required: false
          description: 検索クエリ
          schema:
            type: string
            example: "ハリーポッター"
        - name: page
          in: query
          required: false
          description: ページ指定
          schema:
            type: integer
            example: 2
        - name: sort
          in: query
          required: false
          description: ソート条件
          schema:
            type: string
            enum:
              - popular
              - rating
              - release_date
      responses:
        "200":
          description: 映画のリスト
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  count:
                    type: integer
                  movies:
                    type: array
                    items:
                      $ref: "#/components/schemas/MovieSummary"
  /movies/{movieId}:
    get:
      summary: 映画の詳細情報を返却する
      parameters:
        - name: movieId
          in: path
          required: true
          description: "映画ID"
          schema:
            type: string
            example: "12345"
      responses:
        "200":
          description: 映画の詳細情報とレビュー一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDetail"
  /reviews/{movieId}:
    post:
      summary: 映画に対してレビューを投稿する
      parameters:
        - name: movieId
          in: path
          required: true
          description: "映画ID"
          schema:
            type: string
            example: "12345"
        - in: header
          name: Authorization
          schema:
            type: string
            example: "Bearer xxxxx"
          required: true
          description: アクセストークン
      responses:
        "201":
          description: レビュー投稿成功
        "403":
          description: アクセストークンが無効
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                rating:
                  type: integer

  /auth/signUp:
    post:
      summary: ユーザー登録をする
      responses:
        "201":
          description: ユーザー登録成功
          headers:
            Authorization:
              description: アクセストークン
              schema:
                type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string

  /auth/login:
    post:
      summary: ログインをする
      responses:
        "200":
          description: ログイン成功
          headers:
            Authorization:
              description: アクセストークン
              schema:
                type: string
        "401":
          description: 認証に失敗
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string

components:
  schemas:
    MovieSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        backdrop_url:
          type: string
        rating:
          type: number
          format: double
          minimum: 0
          maximum: 5.00
          multipleOf: 0.01
    MovieDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        overview:
          type: string
        backdrop_url:
          type: string
        poster_url:
          type: string
        release_date:
          type: string
          example: "2024-12-31"
        runtime:
          type: integer
        production_countries:
          type: array
          items:
            type: string
            example: "日本"
        genres:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
        rating:
          type: number
          format: double
          minimum: 0
          maximum: 5.00
          multipleOf: 0.01
        reviews:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              content:
                type: string
              rating:
                type: number
                format: double
                minimum: 0
                maximum: 5.00
                multipleOf: 0.01
              user:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  icon_url:
                    type: string